+++ apps/web/src/features/tags/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./tags-page";
+export * from "./tags-table";
+export * from "./tag-form-dialog";
