+++ apps/web/src/features/tags/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-tags-list";
+export * from "./use-tag-mutations";
