+++ apps/web/src/features/tags/index.ts
@@ -1,0 +1,4 @@
+export * from "./api";
+export * from "./components";
+export * from "./hooks";
+export * from "./types";
