+++ apps/web/src/features/wiki/api/articles.ts
@@ -94,4 +94,3 @@ export type ArticleDetail = ArticleListItem & {
   metaDescription?: string;
   featuredImageUrl?: string;
 };
-
