+++ apps/web/src/features/wiki/components/article-grid-skeleton.tsx
@@ -1,0 +1,26 @@
+import { Card, CardContent, CardHeader } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
+
+export function ArticleGridSkeleton() {
+  return (
+    <div className="grid gap-3 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+      {Array.from({ length: 8 }).map((_, i) => (
+        <Card key={i} className="overflow-hidden">
+          <CardHeader className="pb-3">
+            <Skeleton className="h-4 w-3/4" />
+            <Skeleton className="h-3 w-1/3 mt-2" />
+          </CardHeader>
+          <CardContent className="space-y-3 pt-0">
+            <Skeleton className="h-3 w-full" />
+            <Skeleton className="h-3 w-5/6" />
+            <div className="flex items-center gap-3 pt-2">
+              <Skeleton className="h-4 w-12" />
+              <Skeleton className="h-4 w-12" />
+              <Skeleton className="h-4 w-16" />
+            </div>
+          </CardContent>
+        </Card>
+      ))}
+    </div>
+  );
+}
