+++ apps/web/src/features/wiki/components/article-grid.tsx
@@ -1,0 +1,11 @@
+import { ArticleCard } from "./article-card";
+
+export function ArticleGrid({ articles }: { articles: any[] }) {
+  return (
+    <div className="grid gap-3 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+      {articles.map((a) => (
+        <ArticleCard key={a.id} article={a} />
+      ))}
+    </div>
+  );
+}
