+++ apps/web/src/features/wiki/components/index.ts
@@ -1,7 +1,11 @@
-export { ArticleEditorToolbar } from "./article-editor-toolbar";
-export { ArticleTitleInput } from "./article-title-input";
+export { ArticleCard } from "./article-card";
+export { ArticleGrid } from "./article-grid";
+export { ArticleGridSkeleton } from "./article-grid-skeleton";
+export { EmptyState } from "./empty-state";
+export { NewArticleButton } from "./new-article-button";
 export { ArticleCategorySelect } from "./article-category-select";
-export { ArticleTagsInput } from "./article-tags-input";
-export { ArticleMetadata } from "./article-metadata";
 export { ArticleContentEditor } from "./article-content-editor";
-export type { TagOption } from "./article-tags-input";
+export { ArticleEditorToolbar } from "./article-editor-toolbar";
+export { ArticleMetadata } from "./article-metadata";
+export { ArticleTagsInput } from "./article-tags-input";
+export { ArticleTitleInput } from "./article-title-input";
